refactor: move tRPC files and implement modules directory for feature-based structure
