+++ src/app/(marketing)/page.tsx
@@ -1,7 +1,10 @@
 import { Button } from "@/components/ui/button";
 import { SignInButton, SignedIn, SignedOut, UserButton } from "@clerk/nextjs";
+import { api } from "@/trpc/server";
 
-export default function Home() {
+export default async function Home() {
+  const hello = await api.post.hello({ text: "from tRPC" });
+
   return (
     <div>
       <h1>Home</h1>
@@ -13,6 +16,7 @@ export default function Home() {
           <SignInButton mode="modal">Sign In</SignInButton>
         </Button>
       </SignedOut>
+      <p>{hello ? hello.greeting : "Loading tRPC query..."}</p>
     </div>
   );
 }
