+++ src/app/api/trpc/[trpc]/route.ts
@@ -2,8 +2,8 @@ import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 import { type NextRequest } from "next/server";
 
 import { env } from "@/env";
-import { appRouter } from "@/server/api/root";
-import { createTRPCContext } from "@/server/api/trpc";
+import { appRouter } from "@/trpc/root";
+import { createTRPCContext } from "@/trpc/init";
 
 /**
  * This wraps the `createTRPCContext` helper and provides the required context for the tRPC API when
