+++ src/modules/posts/server/procedures.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { createTRPCRouter, publicProcedure } from "@/server/api/trpc";
+import { createTRPCRouter, publicProcedure } from "@/trpc/init";
 
 export const postRouter = createTRPCRouter({
   hello: publicProcedure
