+++ src/trpc/init.ts
@@ -10,7 +10,7 @@ import { initTRPC } from "@trpc/server";
 import superjson from "superjson";
 import { ZodError } from "zod";
 
-import { db } from "@/server/db";
+import { db } from "@/lib/db";
 
 /**
  * 1. CONTEXT
