+++ src/trpc/root.ts
@@ -1,5 +1,5 @@
-import { postRouter } from "@/server/api/routers/post";
-import { createCallerFactory, createTRPCRouter } from "@/server/api/trpc";
+import { postRouter } from "@/modules/posts/server/procedures";
+import { createCallerFactory, createTRPCRouter } from "@/trpc/init";
 
 /**
  * This is the primary router for your server.
