+++ src/trpc/server.ts
@@ -4,8 +4,8 @@ import { createHydrationHelpers } from "@trpc/react-query/rsc";
 import { headers } from "next/headers";
 import { cache } from "react";
 
-import { createCaller, type AppRouter } from "@/server/api/root";
-import { createTRPCContext } from "@/server/api/trpc";
+import { createCaller, type AppRouter } from "@/trpc/root";
+import { createTRPCContext } from "@/trpc/init";
 import { createQueryClient } from "./query-client";
 
 /**
